#2 Restructure Access Mechanism -Collected previous files into first_attempt package -Created second_attempt package for new system -Created membership package with the relevant files -Created partnership package with the relevant files
Remaining Tasks
-Check to make sure the database mapping is correct
-Fill out volunteer, renter, and guest pieces.
