+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'SignOnServer'
+rootProject.name = 'sign_on_server'
