--- src/main/java/org/powellmakerspace/SignOnServer/SignOnServerApplication.java
@@ -1,13 +1,0 @@
-package org.powellmakerspace.SignOnServer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class SignOnServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SignOnServerApplication.class, args);
-	}
-}
