--- src/main/java/org/powellmakerspace/SignOnServer/exception/ResourceNotFoundException.java
@@ -1,9 +1,0 @@
-package org.powellmakerspace.SignOnServer.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(code = HttpStatus.NOT_FOUND, reason = "Resource Not Found")
-public class ResourceNotFoundException extends Exception {
-
-}
