--- src/main/java/org/powellmakerspace/SignOnServer/models/enums/MembershipType.java
@@ -1,11 +1,0 @@
-package org.powellmakerspace.SignOnServer.models.enums;
-
-public enum MembershipType {
-    INDIVIDUAL, // Added
-    FAMILY, // Added
-    LIBRARY_PASS,
-    PUNCH_PASS, // Added
-    NORTHWEST_COLLEGE,
-    NONMEMBER,
-    VISTA
-}
