--- src/main/java/org/powellmakerspace/SignOnServer/models/enums/VisitPurpose.java
@@ -1,11 +1,0 @@
-package org.powellmakerspace.SignOnServer.models.enums;
-
-public enum VisitPurpose {
-    CLASS,
-    TEACH,
-    VOLUNTEER,
-    WORK_ON_PROJECT,
-    COWORKSPACE,
-    COMMERCIAL_KITCHEN,
-    VISTA
-}
