+++ src/main/java/org/powellmakerspace/sign_on_server/SignOnServerApplication.java
@@ -1,0 +1,13 @@
+package org.powellmakerspace.sign_on_server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class SignOnServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SignOnServerApplication.class, args);
+	}
+}
