+++ src/main/java/org/powellmakerspace/sign_on_server/exception/ResourceNotFoundException.java
@@ -1,0 +1,9 @@
+package org.powellmakerspace.sign_on_server.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(code = HttpStatus.NOT_FOUND, reason = "Resource Not Found")
+public class ResourceNotFoundException extends Exception {
+
+}
