+++ src/main/java/org/powellmakerspace/sign_on_server/models/enums/MembershipType.java
@@ -1,0 +1,11 @@
+package org.powellmakerspace.sign_on_server.models.enums;
+
+public enum MembershipType {
+    INDIVIDUAL, // Added
+    FAMILY, // Added
+    LIBRARY_PASS, // Added
+    PUNCH_PASS, // Added
+    NORTHWEST_COLLEGE,
+    NONMEMBER,
+    VISTA // Kind of added - Needs work
+}
