+++ src/main/java/org/powellmakerspace/sign_on_server/models/enums/VisitPurpose.java
@@ -1,0 +1,11 @@
+package org.powellmakerspace.sign_on_server.models.enums;
+
+public enum VisitPurpose {
+    CLASS,
+    TEACH,
+    VOLUNTEER,
+    WORK_ON_PROJECT,
+    COWORKSPACE,
+    COMMERCIAL_KITCHEN,
+    VISTA
+}
