+++ src/main/java/org/powellmakerspace/sign_on_server/models/membership/second_attempt/AccessMechanism.java
@@ -1,0 +1,6 @@
+package org.powellmakerspace.sign_on_server.models.membership.second_attempt;
+
+public interface AccessMechanism {
+
+    boolean canVisit();
+}
