+++ src/main/java/org/powellmakerspace/sign_on_server/models/membership/second_attempt/membership/Membership.java
@@ -1,0 +1,11 @@
+package org.powellmakerspace.sign_on_server.models.membership.second_attempt.membership;
+
+import org.powellmakerspace.sign_on_server.models.membership.second_attempt.AccessMechanism;
+
+import java.time.LocalDate;
+
+public interface Membership extends AccessMechanism {
+
+    LocalDate getStartDate();
+    LocalDate getEndDate();
+}
