+++ src/main/java/org/powellmakerspace/sign_on_server/models/membership/second_attempt/partnership/PartnerMember.java
@@ -1,0 +1,4 @@
+package org.powellmakerspace.sign_on_server.models.membership.second_attempt.partnership;
+
+public interface PartnerMember {
+}
