--- src/test/java/org/powellmakerspace/SignOnServer/SignOnServerApplicationTests.java
@@ -1,16 +1,0 @@
-package org.powellmakerspace.SignOnServer;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SignOnServerApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
