+++ src/test/java/org/powellmakerspace/sign_on_server/SignOnServerApplicationTests.java
@@ -1,0 +1,16 @@
+package org.powellmakerspace.sign_on_server;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class SignOnServerApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
